1st stage bootloader is completely reversed. Sbot.com is completely reversed. All this one does is reboot the machine. 2nd stage bootloader is mostly understood. However, it loads the executive from disk, then twiddles some of the data in the MZEXE header before jumping *somewhere* in that code. What it's twiddling and where it jumps needs to be researched. However, the executive itself appears to be a real, proper MZEXE file, though it isn't executable in MSDOS.
Stage1 is at byte 0 of the disk image, 1 sector long.
Stage2 is at 1c00 of the image, 4 sectors long.
Sbot.com is a plain old file, that just sits in the FAT.
The executive itself is stored strangely. It isn't stored with the sectors in consecutive order on disk. Please see the 2nd stage bootloader for details.